% Generated by roxygen2 (4.0.2): do not edit by hand
\name{geotopExec}
\alias{geotopExec}
\title{Execution of GEOtop Hydrological Model}
\usage{
geotopExec(param = NULL, bin = "/Users/ecor/local/bin/geotop_zh", simpath,
  inpts.file = "geotop.inpts", runpath = "/Users/ecor/ownCloud/job",
  clean = TRUE, recovery = !clean, getKeywords = NULL,
  data.frame = TRUE, date_field = "Date12.DDMMYYYYhhmm.", intern = FALSE,
  param.soil = TRUE, formatter = "\%04d", paramPrefix = "Header", ...)
}
\arguments{
\item{param}{vector of parameters to set for GEOtop simulation. Actually implemented only for soil type parameters. Deafault is \code{NULL}.}

\item{bin}{binary executable file of GEOtop with full path}

\item{simpath}{directory containg the \code{inpts.file}.}

\item{inpts.file}{GEOtop configuretion \code{*.inpts} file. Default is \code{geotop.inpts}.}

\item{runpath}{directory where to run GEOtop}

\item{clean}{logical value. If it is \code{TRUE} previous simulations and other stuff in \code{runpath} were removed. See \code{\link{file.remove}} for functionality.}

\item{recovery}{logical value. If it is \code{TRUE}, GEOtop simulation are recovered or overwriiten according to the GEOtop settings in \code{inpts.file}.}

\item{getKeywords}{character vector containing the keywords of \code{inpts.file} which can be imported after the GEOtop run. Default is \code{NULL}, nothing is imported.}

\item{data.frame}{logical vaue, see  \code{\link{get.geotop.inpts.keyword.value}}.  Default is \code{TRUE}, is enabled if \code{getKeywords} is not \code{NULL}.}

\item{date_field}{character vaue, see  \code{\link{get.geotop.inpts.keyword.value}}. Default is \code{"Date12.DDMMYYYYhhmm."}, is enabled if \code{getKeywords} is not \code{NULL}.}

\item{intern}{logical value, see \code{\link{system}}.}

\item{param.soil}{logical value If it \code{TRUE} (default) the vaues in \code{param} concerns soil parameters.}

\item{formatter}{character decimal formatter.  See \code{\link{get.geotop.inpts.keyword.value}}. Default is \code{"\%04d"}.}

\item{paramPrefix}{character string. Default is \code{"Header"}. If \code{param.soil==TRUE}, the soil parameters in \code{param} are named with the corrosponding parameter keywords in \code{inpts.file} to which the  string \code{paramPrefix} is attached as a prefix. If \code{paramPrefix} is \code{NA} or \code{NULL} , the names of \code{param} elements must be matched with the corresponimg paramater name in the soil parameter data frame.}

\item{...}{further arguments for \code{\link{get.geotop.inpts.keyword.value}}}
}
\description{
Execution of GEOtop Hydrological Model
}
\details{
In this function, implementation, The parameters entered through \code{param} replace  only the ones of  the first (\code{"0001"}) soil type  declared in the GEOtop simulation directory.
}
\examples{
simpath <- system.file("Muntatschini_pnt_1_225_B2_004",package="geotopOptim")
bin <-   "/Users/ecor/local/bin/geotop_zh"
runpath <- "/Users/ecor/ownCloud/job"

vars <- c("SoilAveragedTempProfileFile",	"SoilLiqWaterPressProfileFile",
"SoilLiqContentProfileFile","SoilIceContentProfileFile")


out <- geotopExec(bin=bin,simpath=simpath,runpath=runpath,
      clean=TRUE,getKeywords=vars,data.frame=TRUE,level=1)

param <- c(N=1.4,Alpha=0.0021,ThetaRes=0.05)
out1 <- geotopExec(param=param,bin=bin,simpath=simpath,
			runpath=runpath,clean=TRUE,getKeywords=vars,
			data.frame=TRUE,level=1,intern=TRUE)
}

