% Generated by roxygen2 (4.0.2): do not edit by hand
\name{geotopPSO}
\alias{geotopPSO}
\alias{geotoplhoat}
\title{GEOtop calibration through Particle Swam Optimization}
\usage{
geotopPSO(fn = geotopGOF, gof.expected.value.for.optim = NA,
  gof.mes = "KGE", weights = "uniform", final.run = TRUE, upper, lower,
  ..., hydroPSOfun = c("hydroPSO", "lhoat"))

geotoplhoat(...)
}
\arguments{
\item{fn}{function to optimize (minimize or maximize). Default is \code{\link{geotopGOF}}. See \code{\link{hydroPSO}}.}

\item{gof.expected.value.for.optim}{expected value for Goodness-of-fit mesure, e.g. 0 or 1. It is used if this function is called by \code{link{geotopPSO}},\code{link{hydroPSO}} or \code{link{optim}}.}

\item{gof.mes}{string(s) containing adopted numerical goodness-of-fit measure. If it is \code{NULL} (Default), all mesasures returned by \code{\link{gof}} are calculated.}

\item{final.run}{logical value. It is \code{TRUE} (default), simulated time series with optimal set of parameteers are added in the list object returned by the function.}

\item{upper,lower}{see  \code{upper} and \code{lowe} in \code{\link{hydroPSO}}}

\item{...}{further arguments for \code{\link{hydroPSO}}.}

\item{hydroPSOfun}{used function name of \code{hydroPSO} package: \code{\link{hydroPSO}} or \code{\link{lhoat}}.}

\item{par}{model parameters. See \code{\link{hydroPSO}},\code{\link{geotopGOF}} and \code{\link{geotopExec}}}
}
\description{
This function performs a calibration or a senisitivity analyisis of
the GEOtop Distributed Hydrological Model through a Particle Swam Optimization or LH-OAT method respectively.
This function is a wrapper of \code{\link{hydroPSO}} or \code{\link{lhoat}}.
The use of either  \code{\link{hydroPSO}} or \code{\link{lhoat}} is declared by the argument \code{hydroPSOfun}.


}
\details{
The function \code{fn}, in case it is different from the default value \code{\link{geotopGOF}} , must always have the arguments \code{gof.mes} and \code{gof.expected.value.for.optim}.
}
\examples{
#'
data(MuntatschiniB2)
## OBSERVATION PROCESSING

obs_SWC <- MuntatschiniB2[str_detect(names(MuntatschiniB2),"SWC")]
zvalues <-  as.numeric(unlist(lapply(X=str_split(names(obs_SWC), pattern="",
				n = Inf),FUN=function (x) {
					out <- as.numeric(x)
				    out <- out[!is.na(out)]
					out <- paste(out,collapse="")
				return(out)
})))
zformatter = "z\%04d"
names(obs_SWC) <- sprintf(zformatter,zvalues)
obs_SWC <- lapply(X=obs_SWC,FUN=function(x){

				if (length(dim(x))>1) {
					max <- apply(X=x,MARGIN=1,FUN=max,na.rm=TRUE)
					min <- apply(X=x,MARGIN=1,FUN=min,na.rm=TRUE)
					mean <- apply(X=x,MARGIN=1,FUN=mean,na.rm=TRUE)
					sd <- apply(X=x,MARGIN=1,FUN=sd,na.rm=TRUE)
				} else {

					mean <- as.vector(x)
				 	max <-  as.vector(x)
					min <-  as.vector(x)
					sd <- NA

				}
				out <- data.frame(min=min,mean=mean,max=max,sd=sd)

				out <- as.zoo(out)
             index(out) <- as.POSIXlt(index(x))

				return(out)

})
###########
###########


simpath <- system.file("Muntatschini_pnt_1_225_B2_004",package="geotopOptim")
bin <-   "/Users/ecor/local/bin/geotop_zh"
runpath <- "/Users/ecor/ownCloud/job"

vars <- "SoilLiqContentProfileFile"

### Use geotopGOF with an internal GEOtop simulation

## create a list with arguments for geotopGOF

x <- param <- c(N=1.4,Alpha=0.0021,ThetaRes=0.05,LateralHydrConductivity=0.021,NormalHydrConductivity=0.021)
upper <- x*3

upper["LateralHydrConductivity"] <- 0.1
upper["NormalHydrConductivity"] <- 0.1

lower <- x/3
lower["N"] <- 1.1
lower["LateralHydrConductivity"] <- 0
lower["NormalHydrConductivity"] <- 0


geotop.model <- list(bin=bin,simpath=simpath,runpath=runpath,
clean=TRUE,variable=vars,data.frame=TRUE,level=1,zformatter=zformatter,intern=TRUE)
control <- list(maxit=4,npart=2) ## Maximim 4 iterations!!

pso <- geotopPSO(par=x,obs=obs_SWC,geotop.model=geotop.model,layer=c("z0020"),gof.mes="KGE",lower=lower,upper=upper,control=control)

lhoat <- geotoplhoat(par=x,obs=obs_SWC,geotop.model=geotop.model,layer=c("z0020"),gof.mes="KGE",lower=lower,upper=upper,control=control)
}
\seealso{
\code{\link{hydroPSO}},\code{\link{gof}}\code{\link{lhoat}}
}

