% Generated by roxygen2 (4.0.2): do not edit by hand
\name{geotopZProfile}
\alias{geotopZProfile}
\title{Execution of GEOtop Hydrological Model and extraction of time series for soil dapth variables}
\usage{
geotopZProfile(..., variable = c("SoilAveragedTempProfileFile",
  "SoilLiqWaterPressProfileFile", "SoilLiqContentProfileFile",
  "SoilIceContentProfileFile"), zformatter = "z\%04d",
  unit = c("centimeters", "millimeters"), geotop_unit = "millimeters")
}
\arguments{
\item{...}{arguments for \code{\link{geotopExec}}}

\item{variable}{name of "Profile' variable to extract. Choose a subset of the default values. It is passed to \code{\link{geotopExec}} through \code{getKeywords} argument.}

\item{zformatter}{decimal formatters for soil layer identification. Default is \code{"z\%04d"}, expressed in centimeters.}

\item{unit}{z coordinate measurement unit. GEOtop values expressed in millimeters which are converted to centimeters by default. Default is \code{c("centimeters","millimeters")}. Otherwise can be the ratio between the unit and one meter.}

\item{geotop_unit}{z coordinate measurement unit used by GEOtop. Default is \code{millimeters}.}
}
\description{
Execution of GEOtop Hydrological Model and extraction of time series for soil dapth variables
}
\examples{
simpath <- system.file("Muntatschini_pnt_1_225_B2_004",package="geotopOptim")
bin <-   "/Users/ecor/local/bin/geotop_zh"
runpath <- "/Users/ecor/ownCloud/job"

vars <- c("SoilAveragedTempProfileFile",	"SoilLiqWaterPressProfileFile",
"SoilLiqContentProfileFile","SoilIceContentProfileFile")

vars <- "SoilLiqContentProfileFile"

out <- geotopZProfile(bin=bin,simpath=simpath,runpath=runpath,
			clean=TRUE,variable=vars,data.frame=TRUE,level=1)
}
\seealso{
\code{\link{geotopExec}}
}

